refactor: replace utils.CountWithTimeout with db.CountWithTimeout across repositories
- Updated multiple repository files to replace calls to utils.CountWithTimeout with the new db.CountWithTimeout function, ensuring consistent timeout handling for count queries.
- Removed the .air.toml file from the api directory and added Go binaries to .gitignore to streamline the build process.
- Introduced a new count.go file in the db package to implement the CountWithTimeout function, enhancing query execution with timeout management.
