+++ api/repositories/block.go
@@ -7,7 +7,6 @@ import (
 	"gorm.io/gorm/clause"
 
 	"github.com/initia-labs/core-indexer/api/dto"
-	"github.com/initia-labs/core-indexer/api/utils"
 	"github.com/initia-labs/core-indexer/pkg/db"
 	"github.com/initia-labs/core-indexer/pkg/logger"
 )
@@ -228,7 +227,7 @@ func (r *BlockRepository) GetBlockTxs(pagination dto.PaginationQuery, height int
 
 	if pagination.CountTotal {
 		var err error
-		total, err = utils.CountWithTimeout(r.db.Model(&db.Transaction{}).Joins("LEFT JOIN blocks ON blocks.height = transactions.block_height").Where("blocks.height = ?", height), r.countQueryTimeout)
+		total, err = db.CountWithTimeout(r.db.Model(&db.Transaction{}).Joins("LEFT JOIN blocks ON blocks.height = transactions.block_height").Where("blocks.height = ?", height), r.countQueryTimeout)
 		if err != nil {
 			logger.Get().Error().Err(err).Msg("Failed to get total block transaction count")
 			return nil, 0, err
